Add successful and skipped tests to payload.
This commit adds a BuildCompleteListener that listens on the PostBuildCompletedEvent. The testcases provided in the BuildContext get saved into a Map and can be accessed from the ChainCompletedNotification.
See the comment above the Map cachedTestResults in the class ServerNotificationRecipient for more information.

Updated Bamboo-api-version.

Updated sample payload shown on README.